requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
clarifai-grpc==9.0.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.38.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
protobuf==4.21.12
pyexiv2==2.8.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.14
